[#22] 로그인 구현 - 로그인 서비스 인터페이스를 활용한 분리(확장성 고려)
